Updated Langchain 2024 - KrishNayak/requirements.txt
langchain
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community

langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
PyPDF2
chromadb
faiss-cpu
langchain_huggingface
langchain-huggingface-hub
wikipedia
arxiv
langchainhub


tqdm
nltk
groq
pandas
cassio
requests
python-docx
unstructured
beautifulsoup4
langchain_groq
google-generativeai
sentence-transformers
sentence_transformers
langchain_google_genai


torch 
transformers 
pillow